handle ェ
