+++ src/Enums/Youon.cs
@@ -4,6 +4,7 @@
 	{
 		Ya,
 		Yu,
+		Ye,
 		Yo
 	}
 }
